+++ src/main/java/org/folio/cataloging/business/cataloguing/authority/AuthorityHeaderFieldHelper.java
@@ -20,7 +20,7 @@ public class AuthorityHeaderFieldHelper extends HeaderFieldHelper {
 	/* (non-Javadoc)
 	 * @see HeaderField#getCategory()
 	 */
-	public short getCategory() {
+	public int getCategory() {
 		return 1;
 	}
