+++ src/main/java/org/folio/cataloging/business/cataloguing/bibliographic/SubjectAccessPoint.java
@@ -4,7 +4,6 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.folio.cataloging.business.cataloguing.common.AccessPoint;
 import org.folio.cataloging.business.cataloguing.common.OrderedTag;
-import org.folio.cataloging.business.common.CorrelationValues;
 import org.folio.cataloging.business.common.DataAccessException;
 import org.folio.cataloging.business.common.Defaults;
 import org.folio.cataloging.business.descriptor.Descriptor;
@@ -12,6 +11,7 @@ import org.folio.cataloging.dao.DAOBibliographicCorrelation;
 import org.folio.cataloging.dao.DAODescriptor;
 import org.folio.cataloging.dao.DAOSubjectDescriptor;
 import org.folio.cataloging.dao.persistence.*;
+import org.folio.cataloging.shared.CorrelationValues;
 import org.folio.cataloging.util.StringText;
 
 import java.util.ArrayList;
@@ -43,7 +43,7 @@ public class SubjectAccessPoint extends BibliographicAccessPoint implements Orde
 		setFunctionCode(Defaults.getShort("subjectAccessPoint.functionCode"));
 	}
 
-	public short getFunctionCode() {
+	public int getFunctionCode() {
 		return functionCode;
 	}
 
@@ -121,9 +121,8 @@ public class SubjectAccessPoint extends BibliographicAccessPoint implements Orde
 		} catch (MarcCorrelationException e) {
 			try {
 				CorrelationValues v = getCorrelationValues();
-				short v2 = new DAOBibliographicCorrelation()
-				.getFirstAllowedValue2(getCategory(), v.getValue(1), v
-						.getValue(3));
+				int v2 = new DAOBibliographicCorrelation()
+				.getFirstAllowedValue2(getCategory(), v.getValue(1), v.getValue(3));
 				setFunctionCode(v2);
 			} catch (DataAccessException e1) {
 				setFunctionCode((short)-1);
@@ -210,7 +209,7 @@ public class SubjectAccessPoint extends BibliographicAccessPoint implements Orde
 	 * 
 	 * @see librisuite.business.cataloguing.bibliographic.Tag#getCategory()
 	 */
-	public short getCategory() {
+	public int getCategory() {
 		return 4;
 	}
