+++ src/main/java/org/folio/cataloging/dao/persistence/CNTL_NBR.java
@@ -10,15 +10,14 @@ package org.folio.cataloging.dao.persistence;
 import org.folio.cataloging.business.cataloguing.authority.AuthorityControlNumberAccessPoint;
 import org.folio.cataloging.business.cataloguing.bibliographic.ControlNumberAccessPoint;
 import org.folio.cataloging.business.common.ConfigHandler;
-import org.folio.cataloging.business.common.CorrelationValues;
 import org.folio.cataloging.business.common.DataAccessException;
 import org.folio.cataloging.business.common.Defaults;
-import org.folio.cataloging.dao.DAOControlNumberDescriptor;
 import org.folio.cataloging.business.descriptor.Descriptor;
 import org.folio.cataloging.business.descriptor.SortFormParameters;
+import org.folio.cataloging.dao.DAOControlNumberDescriptor;
 import org.folio.cataloging.dao.DAOIndexList;
-
 import org.folio.cataloging.dao.common.HibernateUtil;
+import org.folio.cataloging.shared.CorrelationValues;
 
 /**
  * @author paulm
@@ -27,7 +26,7 @@ import org.folio.cataloging.dao.common.HibernateUtil;
  */
 public class CNTL_NBR extends Descriptor {
 	private static final long serialVersionUID = 1L;
-	private short typeCode;
+	private int typeCode;
 	private ConfigHandler configHandler =ConfigHandler.getInstance();
 	/**
 	 * Class constructor
@@ -126,7 +125,7 @@ public class CNTL_NBR extends Descriptor {
 	/* (non-Javadoc)
 	 * @see librisuite.hibernate.Descriptor#getCategory()
 	 */
-	public short getCategory() {
+	public int getCategory() {
 		return 19;
 	}
 
@@ -134,7 +133,7 @@ public class CNTL_NBR extends Descriptor {
 	 * 
 	 * @since 1.0
 	 */
-	public short getTypeCode() {
+	public int getTypeCode() {
 		return typeCode;
 	}
 
@@ -142,7 +141,7 @@ public class CNTL_NBR extends Descriptor {
 	 * 
 	 * @since 1.0
 	 */
-	public void setTypeCode(short s) {
+	public void setTypeCode(int s) {
 		typeCode = s;
 	}
