+++ src/main/java/org/folio/cataloging/dao/persistence/NME_TTL_HDG.java
@@ -8,14 +8,13 @@
 package org.folio.cataloging.dao.persistence;
 
 import org.folio.cataloging.business.cataloguing.bibliographic.NameTitleAccessPoint;
-import org.folio.cataloging.business.common.CorrelationValues;
-import org.folio.cataloging.dao.DAONameTitleDescriptor;
 import org.folio.cataloging.business.descriptor.Descriptor;
 import org.folio.cataloging.business.descriptor.SortFormParameters;
-
+import org.folio.cataloging.dao.DAONameTitleDescriptor;
 import org.folio.cataloging.dao.common.HibernateUtil;
-import org.folio.cataloging.util.StringText;
 import org.folio.cataloging.model.Subfield;
+import org.folio.cataloging.shared.CorrelationValues;
+import org.folio.cataloging.util.StringText;
 
 /**
  * @author paulm
@@ -62,7 +61,7 @@ public class NME_TTL_HDG extends Descriptor {
 	/* (non-Javadoc)
 	 * @see Descriptor#getCategory()
 	 */
-	public short getCategory() {
+	public int getCategory() {
 		return 11;
 	}
